refactor: 패키지 구조 수정하라
- validator 객체는 domain 을 위한 객체이기 떄문에 domain 아래로 이동한다
