+++ src/main/java/lotto/domain/validator/PriceValidator.java
@@ -1,4 +1,4 @@
-package lotto.validator;
+package lotto.domain.validator;
 
 public interface PriceValidator {
