+++ src/test/java/lotto/validator/InvalidWinningNumberSizeValidatorTest.java
@@ -2,8 +2,9 @@ package lotto.validator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.validator.WinningNumbersValidator;
 import lotto.exception.ExceptionType;
-import lotto.validator.impl.InvalidWinningNumberSizeValidator;
+import lotto.domain.validator.impl.InvalidWinningNumberSizeValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
