+++ src/test/java/lotto/validator/WinningNumberValidatorGroupTest.java
@@ -3,6 +3,7 @@ package lotto.validator;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import lotto.domain.validator.WinningNumberValidatorGroup;
 import lotto.exception.ExceptionType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
